feat(cpps): Send order states to MF agent
Previously, only the AMR was aware of the current order states.
Updates were not shared with the logical MF agent
Update logging was (partially) done from the logical AMR itself.

Now the update is sent to the MF agent through unicast messages.
The MF logger contains a logger "algorithm" that processes these
messages.

This must be a algorithm because of the design of LogicalAgent.
This is related to issue #79